refactor(recruitment): move test fixture
